api_yamdb/requirements.txt
asgiref== 3.6.0
Django== 3.2
django-filter== 22.1
djangorestframework==3.12.4
djangorestframework-simplejwt== 5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
python-dotenv==1.0.0
requests==2.28.2
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4

